Refactor state.rb that separate ResolvedConflict class from lrama/state.rb
